Migrate app lifecycle from AppDelegate to SceneDelegate
Adopt the UIKit scene lifecycle: add a UIApplicationSceneManifest to
Info.plist (replacing UIMainStoryboardFile) and introduce SceneDelegate
as the window owner.

Move everything that arrives through a UIWindowScene off AppDelegate and
onto SceneDelegate / SceneDelegateViewModel: URL opens, continued user
activities, shortcut items, email-verification redirects, and the
URLSessionTaskDelegate used to resolve email-click links. Push
notification and Braze deep links stay on AppDelegateViewModel.

Both view models resolve Navigation values through a new shared
DeepLinkNavigationRouter so the routing logic isn't duplicated.

Once an app adopts scenes, UIKit stops calling the UIApplicationDelegate
foreground/background/active callbacks, so SceneDelegate implements the
scene equivalents and forwards them to AppDelegateViewModel, which still
owns the app-scoped work they drive (user and config refresh, icon badge
reset, app-tracking authorization). SceneDelegateViewModel drops the
foreground event UIKit sends during initial scene connection, which
would otherwise duplicate applicationDidFinishLaunching.

AppDelegate.rootTabBarController now resolves through the connected
scenes' windows, matching on the root view controller's type rather than
taking windows.first, whose order isn't guaranteed and which can be a
system window such as the keyboard's UITextEffectsWindow.
