+++ Kickstarter-Framework/Sources/Kickstarter-Framework-iOSTests/Kickstarter-iOS/AppDelegateViewModelTests.swift
@@ -18,9 +18,6 @@ final class AppDelegateViewModelTests: TestCase {
   private let configureFirebase = TestObserver<(), Never>()
   private let configureSegmentWithBraze = TestObserver<String, Never>()
   private let didAcceptReceivingRemoteNotifications = TestObserver<(), Never>()
-  private let emailVerificationCompletedMessage = TestObserver<String, Never>()
-  private let emailVerificationCompletedSuccess = TestObserver<Bool, Never>()
-  private let findRedirectUrl = TestObserver<URL, Never>()
   private let forceLogout = TestObserver<(), Never>()
   private let goToActivity = TestObserver<(), Never>()
   private let goToDiscovery = TestObserver<DiscoveryParams?, Never>()
@@ -62,11 +59,6 @@ final class AppDelegateViewModelTests: TestCase {
     self.vm.outputs.applicationIconBadgeNumber.observe(self.applicationIconBadgeNumber.observer)
     self.vm.outputs.configureFirebase.observe(self.configureFirebase.observer)
     self.vm.outputs.configureSegmentWithBraze.observe(self.configureSegmentWithBraze.observer)
-    self.vm.outputs.emailVerificationCompleted.map(first)
-      .observe(self.emailVerificationCompletedMessage.observer)
-    self.vm.outputs.emailVerificationCompleted.map(second)
-      .observe(self.emailVerificationCompletedSuccess.observer)
-    self.vm.outputs.findRedirectUrl.observe(self.findRedirectUrl.observer)
     self.vm.outputs.forceLogout.observe(self.forceLogout.observer)
     self.vm.outputs.goToActivity.observe(self.goToActivity.observer)
     self.vm.outputs.goToDiscovery.observe(self.goToDiscovery.observer)
@@ -298,619 +290,6 @@ final class AppDelegateViewModelTests: TestCase {
         .ksr_configUpdated,
         .ksr_configUpdated
       ])
-    }
-  }
-
-  func testPresentViewController() {
-    let apiService = MockService(
-      fetchProjectResult: .success(.template),
-      fetchUpdateResponse: .template
-    )
-    withEnvironment(apiService: apiService) {
-      let rootUrl = "https://www.kickstarter.com/"
-
-      self.vm.inputs.applicationDidFinishLaunching(
-        application: UIApplication.shared,
-        launchOptions: [:]
-      )
-
-      self.presentViewController.assertValues([])
-
-      let projectUrl = rootUrl + "projects/tequila/help-me-transform-this-pile-of-wood"
-      var result = self.vm.inputs.applicationOpenUrl(
-        application: UIApplication.shared,
-        url: URL(string: projectUrl)!,
-        options: [:]
-      )
-      XCTAssertTrue(result)
-
-      self.presentViewController.assertValues([1])
-
-      let commentsUrl = projectUrl + "/comments"
-      result = self.vm.inputs.applicationOpenUrl(
-        application: UIApplication.shared,
-        url: URL(string: commentsUrl)!,
-        options: [:]
-      )
-      XCTAssertTrue(result)
-
-      self.presentViewController.assertValues([1, 2])
-
-      let updatesUrl = projectUrl + "/posts"
-      result = self.vm.inputs.applicationOpenUrl(
-        application: UIApplication.shared,
-        url: URL(string: updatesUrl)!,
-        options: [:]
-      )
-      XCTAssertTrue(result)
-
-      self.presentViewController.assertValues([1, 2, 2])
-
-      let updateUrl = projectUrl + "/posts/1399396"
-      result = self.vm.inputs.applicationOpenUrl(
-        application: UIApplication.shared,
-        url: URL(string: updateUrl)!,
-        options: [:]
-      )
-      XCTAssertTrue(result)
-
-      self.presentViewController.assertValues([1, 2, 2, 2])
-
-      let updateCommentsUrl = updateUrl + "/comments"
-      result = self.vm.inputs.applicationOpenUrl(
-        application: UIApplication.shared,
-        url: URL(string: updateCommentsUrl)!,
-        options: [:]
-      )
-      XCTAssertTrue(result)
-
-      self.presentViewController.assertValues([1, 2, 2, 2, 3])
-
-      let faqUrl = projectUrl + "/faqs"
-      result = self.vm.inputs.applicationOpenUrl(
-        application: UIApplication.shared,
-        url: URL(string: faqUrl)!,
-        options: [:]
-      )
-      XCTAssertTrue(result)
-
-      self.presentViewController.assertValues([1, 2, 2, 2, 3, 1])
-    }
-  }
-
-  func testPresentViewController_ProjectPreviewLink_DisplayPrelaunch_True() {
-    let project = Project.template
-      |> Project.lens.displayPrelaunch .~ true
-
-    let apiService = MockService(fetchProjectResult: .success(project))
-    withEnvironment(apiService: apiService) {
-      let rootUrl = "https://www.kickstarter.com/"
-
-      self.vm.inputs.applicationDidFinishLaunching(
-        application: UIApplication.shared,
-        launchOptions: [:]
-      )
-
-      self.presentViewController.assertValues([])
-
-      let projectUrl = rootUrl + "projects/tequila/help-me-transform-this-pile-of-wood"
-      let result = self.vm.inputs.applicationOpenUrl(
-        application: UIApplication.shared,
-        url: URL(string: projectUrl)!,
-        options: [:]
-      )
-      XCTAssertTrue(result)
-
-      self.presentViewController.assertValues([1])
-      self.goToMobileSafari.assertValues([])
-    }
-  }
-
-  func testPresentViewController_ProjectPreviewLink_DisplayPrelaunch_False() {
-    let project = Project.template
-      |> Project.lens.displayPrelaunch .~ false
-
-    let apiService = MockService(fetchProjectResult: .success(project))
-    withEnvironment(apiService: apiService) {
-      let rootUrl = "https://www.kickstarter.com/"
-
-      self.vm.inputs.applicationDidFinishLaunching(
-        application: UIApplication.shared,
-        launchOptions: [:]
-      )
-
-      self.presentViewController.assertValues([])
-
-      let projectUrl = rootUrl + "projects/tequila/help-me-transform-this-pile-of-wood"
-      let result = self.vm.inputs.applicationOpenUrl(
-        application: UIApplication.shared,
-        url: URL(string: projectUrl)!,
-        options: [:]
-      )
-      XCTAssertTrue(result)
-
-      self.presentViewController.assertValues([1])
-      self.goToMobileSafari.assertValues([])
-    }
-  }
-
-  func testPresentViewController_ProjectPreviewLink_DisplayPrelaunch_Nil() {
-    let project = Project.template
-      |> Project.lens.displayPrelaunch .~ nil
-
-    let apiService = MockService(fetchProjectResult: .success(project))
-    withEnvironment(apiService: apiService) {
-      let rootUrl = "https://www.kickstarter.com/"
-
-      self.vm.inputs.applicationDidFinishLaunching(
-        application: UIApplication.shared,
-        launchOptions: [:]
-      )
-
-      self.presentViewController.assertValues([])
-
-      let projectUrl = rootUrl + "projects/tequila/help-me-transform-this-pile-of-wood"
-      let result = self.vm.inputs.applicationOpenUrl(
-        application: UIApplication.shared,
-        url: URL(string: projectUrl)!,
-        options: [:]
-      )
-      XCTAssertTrue(result)
-
-      self.presentViewController.assertValues([1])
-      self.goToMobileSafari.assertValues([])
-    }
-  }
-
-  func testPresentViewController_ProjectCommentThread_Success() {
-    self.vm.inputs.applicationDidFinishLaunching(
-      application: UIApplication.shared,
-      launchOptions: [:]
-    )
-
-    withEnvironment(apiService: MockService(fetchCommentRepliesEnvelopeResult: .success(
-      CommentRepliesEnvelope
-        .successfulRepliesTemplate
-    ), fetchProjectResult: .success(.template))) {
-      let url =
-        "https://\(AppEnvironment.current.apiService.serverConfig.webBaseUrl.host ?? "")/projects/fjorden/fjorden-iphone-photography-reinvented/comments?comment=Q29tbWVudC0zMzY0OTg0MQ%3D%3D"
-
-      let result = self.vm.inputs.applicationOpenUrl(
-        application: UIApplication.shared,
-        url: URL(string: url)!,
-        options: [:]
-      )
-      XCTAssertTrue(result)
-
-      self.presentViewController.assertValues([3])
-    }
-  }
-
-  func testPresentViewController_ProjectCommentThread_Reply_Success() {
-    self.vm.inputs.applicationDidFinishLaunching(
-      application: UIApplication.shared,
-      launchOptions: [:]
-    )
-
-    withEnvironment(apiService: MockService(fetchCommentRepliesEnvelopeResult: .success(
-      CommentRepliesEnvelope
-        .successfulRepliesTemplate
-    ), fetchProjectResult: .success(.template))) {
-      let url =
-        "https://\(AppEnvironment.current.apiService.serverConfig.webBaseUrl.host ?? "")/projects/fjorden/fjorden-iphone-photography-reinvented/comments?comment=Q29tbWVudC0zMzY0OTg0MQ%3D%3D&reply=deadbeef"
-
-      let result = self.vm.inputs.applicationOpenUrl(
-        application: UIApplication.shared,
-        url: URL(string: url)!,
-        options: [:]
-      )
-      XCTAssertTrue(result)
-
-      self.presentViewController.assertValues([3])
-    }
-  }
-
-  func testPresentViewController_UpdateCommentThread_Success() {
-    self.vm.inputs.applicationDidFinishLaunching(
-      application: UIApplication.shared,
-      launchOptions: [:]
-    )
-
-    withEnvironment(apiService: MockService(fetchCommentRepliesEnvelopeResult: .success(
-      CommentRepliesEnvelope
-        .successfulRepliesTemplate
-    ), fetchProjectResult: .success(.template))) {
-      let url =
-        "https://\(AppEnvironment.current.apiService.serverConfig.webBaseUrl.host ?? "")/projects/fjorden/fjorden-iphone-photography-reinvented/posts/3254626/comments?comment=Q29tbWVudC0zMzY0OTg0MQ%3D%3D"
-
-      let result = self.vm.inputs.applicationOpenUrl(
-        application: UIApplication.shared,
-        url: URL(string: url)!,
-        options: [:]
-      )
-      XCTAssertTrue(result)
-
-      self.presentViewController.assertValues([4])
-    }
-  }
-
-  func testPresentViewController_UpdateCommentThread_Reply_Success() {
-    self.vm.inputs.applicationDidFinishLaunching(
-      application: UIApplication.shared,
-      launchOptions: [:]
-    )
-
-    withEnvironment(apiService: MockService(fetchCommentRepliesEnvelopeResult: .success(
-      CommentRepliesEnvelope
-        .successfulRepliesTemplate
-    ), fetchProjectResult: .success(.template))) {
-      let url =
-        "https://\(AppEnvironment.current.apiService.serverConfig.webBaseUrl.host ?? "")/projects/fjorden/fjorden-iphone-photography-reinvented/posts/3254626/comments?comment=Q29tbWVudC0zMzY0OTg0MQ%3D%3D&reply=deadbeef"
-
-      let result = self.vm.inputs.applicationOpenUrl(
-        application: UIApplication.shared,
-        url: URL(string: url)!,
-        options: [:]
-      )
-      XCTAssertTrue(result)
-
-      self.presentViewController.assertValues([4])
-    }
-  }
-
-  func testGoToActivity() {
-    self.vm.inputs.applicationDidFinishLaunching(
-      application: UIApplication.shared,
-      launchOptions: [:]
-    )
-
-    self.goToActivity.assertValueCount(0)
-
-    let result = self.vm.inputs.applicationOpenUrl(
-      application: UIApplication.shared,
-      url: URL(string: "https://www.kickstarter.com/activity")!,
-      options: [:]
-    )
-    XCTAssertTrue(result)
-
-    self.goToActivity.assertValueCount(1)
-  }
-
-  func testGoToDiscovery() {
-    self.vm.inputs.applicationDidFinishLaunching(
-      application: UIApplication.shared,
-      launchOptions: [:]
-    )
-
-    self.goToDiscovery.assertValues([])
-
-    let result = self.vm.inputs.applicationOpenUrl(
-      application: UIApplication.shared,
-      url: URL(string: "https://www.kickstarter.com/discover?sort=newest")!,
-      options: [:]
-    )
-    XCTAssertTrue(result)
-
-    let params = .defaults
-      |> DiscoveryParams.lens.sort .~ .newest
-    self.goToDiscovery.assertValues([params])
-  }
-
-  func testGoToDiscovery_NoParams() {
-    self.vm.inputs.applicationDidFinishLaunching(
-      application: UIApplication.shared,
-      launchOptions: [:]
-    )
-
-    self.goToDiscovery.assertValues([])
-
-    let result = self.vm.inputs.applicationOpenUrl(
-      application: UIApplication.shared,
-      url: URL(string: "https://www.kickstarter.com/discover")!,
-      options: [:]
-    )
-    XCTAssertTrue(result)
-
-    self.goToDiscovery.assertValues([nil])
-  }
-
-  func testGoToDiscoveryWithCategoryName_ValidCategoryName_RoutesToCategory() {
-    let mockService = MockService(fetchGraphCategoriesResult: .success(defaultRootCategoriesTemplate))
-
-    withEnvironment(apiService: mockService) {
-      self.vm.inputs.applicationDidFinishLaunching(
-        application: UIApplication.shared,
-        launchOptions: [:]
-      )
-
-      self.goToDiscovery.assertValues([])
-
-      let url = URL(string: "https://www.kickstarter.com/discover/categories/art")!
-      let result = self.vm.inputs.applicationOpenUrl(
-        application: UIApplication.shared,
-        url: url,
-        options: [:]
-      )
-      XCTAssertTrue(result)
-
-      self.scheduler.advance()
-
-      let params = .defaults |> DiscoveryParams.lens.category .~ .art
-      self.goToDiscovery.assertValues([params])
-    }
-  }
-
-  func testGoToDiscoveryWithCategoryName_InvalidCategoryName_DoesNotRouteToAnyCategory() {
-    let mockService = MockService(fetchGraphCategoriesResult: .success(defaultRootCategoriesTemplate))
-
-    withEnvironment(apiService: mockService) {
-      self.vm.inputs.applicationDidFinishLaunching(
-        application: UIApplication.shared,
-        launchOptions: [:]
-      )
-
-      self.goToDiscovery.assertValues([])
-
-      let url = URL(string: "https://www.kickstarter.com/discover/categories/random")!
-      let result = self.vm.inputs.applicationOpenUrl(
-        application: UIApplication.shared,
-        url: url,
-        options: [:]
-      )
-      XCTAssertTrue(result)
-
-      self.scheduler.advance()
-
-      let params = .defaults |> DiscoveryParams.lens.category .~ .none
-      self.goToDiscovery.assertValues([params])
-    }
-  }
-
-  func testGoToDiscoveryWithSubcategoryName_ValidSubcategoryName_RoutesToSubcategory() {
-    let gamesTemplate = RootCategoriesEnvelope.template
-      |> RootCategoriesEnvelope.lens.categories .~ [
-        .art,
-        .filmAndVideo,
-        .illustration,
-        .documentary,
-        .games
-      ]
-
-    let mockService = MockService(fetchGraphCategoriesResult: .success(gamesTemplate))
-
-    withEnvironment(apiService: mockService) {
-      self.vm.inputs.applicationDidFinishLaunching(
-        application: UIApplication.shared,
-        launchOptions: [:]
-      )
-
-      self.goToDiscovery.assertValues([])
-
-      let url = URL(string: "https://www.kickstarter.com/discover/categories/games/tabletop%20games")!
-      let result = self.vm.inputs.applicationOpenUrl(
-        application: UIApplication.shared,
-        url: url,
-        options: [:]
-      )
-      XCTAssertTrue(result)
-
-      self.scheduler.advance()
-
-      let params = .defaults |> DiscoveryParams.lens.category .~ .tabletopGames
-      self.goToDiscovery.assertValues([params])
-    }
-  }
-
-  func testGoToDiscoveryWithSubcategoryName_InvalidSubcategoryName_RoutesToCategory() {
-    let gamesTemplate = RootCategoriesEnvelope.template
-      |> RootCategoriesEnvelope.lens.categories .~ [
-        .art,
-        .filmAndVideo,
-        .illustration,
-        .documentary,
-        .games
-      ]
-
-    let mockService = MockService(fetchGraphCategoriesResult: .success(gamesTemplate))
-
-    withEnvironment(apiService: mockService) {
-      self.vm.inputs.applicationDidFinishLaunching(
-        application: UIApplication.shared,
-        launchOptions: [:]
-      )
-
-      self.goToDiscovery.assertValues([])
-
-      let url = URL(string: "https://www.kickstarter.com/discover/categories/games/tabletopgames")!
-      let result = self.vm.inputs.applicationOpenUrl(
-        application: UIApplication.shared,
-        url: url,
-        options: [:]
-      )
-      XCTAssertTrue(result)
-
-      self.scheduler.advance()
-
-      let params = .defaults |> DiscoveryParams.lens.category .~ .games
-      self.goToDiscovery.assertValues([params])
-    }
-  }
-
-  func testGoToDiscoveryWithCategoryId_ValidCategoryId_RoutesToCategory() {
-    let mockService = MockService(fetchGraphCategoriesResult: .success(defaultRootCategoriesTemplate))
-
-    withEnvironment(apiService: mockService) {
-      self.vm.inputs.applicationDidFinishLaunching(
-        application: UIApplication.shared,
-        launchOptions: [:]
-      )
-
-      self.goToDiscovery.assertValues([])
-
-      let url =
-        URL(
-          string: "https://www.kickstarter.com/discover/advanced?category_id=1&sort=magic&seed=2714369&page=1"
-        )!
-      let result = self.vm.inputs.applicationOpenUrl(
-        application: UIApplication.shared,
-        url: url,
-        options: [:]
-      )
-      XCTAssertTrue(result)
-
-      self.scheduler.advance()
-
-      let params = .defaults |> DiscoveryParams.lens.category .~ .art
-        |> DiscoveryParams.lens.sort .~ .magic
-        |> DiscoveryParams.lens.seed .~ 2_714_369
-        |> DiscoveryParams.lens.page .~ 1
-      self.goToDiscovery.assertValues([params])
-    }
-  }
-
-  func testGoToDiscoveryWithCategoryId_InvalidCategoryOrSubcategoryId_DoesNotRouteToAnyCategory() {
-    let mockService = MockService(fetchGraphCategoriesResult: .success(defaultRootCategoriesTemplate))
-
-    withEnvironment(apiService: mockService) {
-      self.vm.inputs.applicationDidFinishLaunching(
-        application: UIApplication.shared,
-        launchOptions: [:]
-      )
-
-      self.goToDiscovery.assertValues([])
-
-      let url =
-        URL(
-          string: "https://www.kickstarter.com/discover/advanced?category_id=9999&sort=magic&seed=2714369&page=1"
-        )!
-      let result = self.vm.inputs.applicationOpenUrl(
-        application: UIApplication.shared,
-        url: url,
-        options: [:]
-      )
-      XCTAssertTrue(result)
-
-      self.scheduler.advance()
-
-      let params = .defaults |> DiscoveryParams.lens.category .~ .none
-        |> DiscoveryParams.lens.sort .~ .magic
-        |> DiscoveryParams.lens.seed .~ 2_714_369
-        |> DiscoveryParams.lens.page .~ 1
-      self.goToDiscovery.assertValues([params])
-    }
-  }
-
-  func testGoToDiscoveryWithSubcategoryId_ValidSubcategoryId_RoutesToSubcategory() {
-    let gamesTemplate = RootCategoriesEnvelope.template
-      |> RootCategoriesEnvelope.lens.categories .~ [
-        .art,
-        .filmAndVideo,
-        .illustration,
-        .documentary,
-        .games
-      ]
-
-    let mockService = MockService(fetchGraphCategoriesResult: .success(gamesTemplate))
-
-    withEnvironment(apiService: mockService) {
-      self.vm.inputs.applicationDidFinishLaunching(
-        application: UIApplication.shared,
-        launchOptions: [:]
-      )
-
-      self.goToDiscovery.assertValues([])
-
-      let url =
-        URL(
-          string: "https://www.kickstarter.com/discover/advanced?category_id=34&sort=magic&seed=2714369&page=1"
-        )!
-      let result = self.vm.inputs.applicationOpenUrl(
-        application: UIApplication.shared,
-        url: url,
-        options: [:]
-      )
-      XCTAssertTrue(result)
-
-      self.scheduler.advance()
-
-      let params = .defaults |> DiscoveryParams.lens.category .~ .tabletopGames
-        |> DiscoveryParams.lens.sort .~ .magic
-        |> DiscoveryParams.lens.seed .~ 2_714_369
-        |> DiscoveryParams.lens.page .~ 1
-      self.goToDiscovery.assertValues([params])
-    }
-  }
-
-  func testGoToLogin() {
-    self.vm.inputs.applicationDidFinishLaunching(
-      application: UIApplication.shared,
-      launchOptions: [:]
-    )
-
-    self.goToLoginWithIntent.assertValueCount(0)
-
-    let result = self.vm.inputs.applicationOpenUrl(
-      application: UIApplication.shared,
-      url: URL(string: "https://www.kickstarter.com/authorize")!,
-      options: [:]
-    )
-    XCTAssertTrue(result)
-
-    self.goToLoginWithIntent.assertValueCount(1)
-  }
-
-  func testGoToProfile() {
-    self.vm.inputs.applicationDidFinishLaunching(
-      application: UIApplication.shared,
-      launchOptions: [:]
-    )
-
-    self.goToProfile.assertValueCount(0)
-
-    let result = self.vm.inputs.applicationOpenUrl(
-      application: UIApplication.shared,
-      url: URL(string: "https://www.kickstarter.com/profile/me")!,
-      options: [:]
-    )
-    XCTAssertTrue(result)
-
-    self.goToProfile.assertValueCount(1)
-  }
-
-  func testGoToSearch() {
-    self.vm.inputs.applicationDidFinishLaunching(
-      application: UIApplication.shared,
-      launchOptions: [:]
-    )
-
-    self.goToSearch.assertValueCount(0)
-
-    let result = self.vm.inputs.applicationOpenUrl(
-      application: UIApplication.shared,
-      url: URL(string: "https://www.kickstarter.com/search")!,
-      options: [:]
-    )
-    XCTAssertTrue(result)
-
-    self.goToSearch.assertValueCount(1)
-  }
-
-  func testDeeplink_IsActivated_Success() {
-    self.vm.inputs.applicationDidFinishLaunching(
-      application: UIApplication.shared,
-      launchOptions: [:]
-    )
-
-    withEnvironment(currentUser: nil) {
-      let result = self.vm.inputs.applicationOpenUrl(
-        application: UIApplication.shared,
-        url: URL(string: "https://www.kickstarter.com/search")!,
-        options: [:]
-      )
-
-      XCTAssertTrue(result)
-
-      self.goToSearch.assertValueCount(1)
     }
   }
 
@@ -1227,50 +606,6 @@ final class AppDelegateViewModelTests: TestCase {
     }
   }
 
-  func testContinueUserActivity_ValidActivity() {
-    let userActivity = NSUserActivity(activityType: NSUserActivityTypeBrowsingWeb)
-    userActivity.webpageURL = URL(string: "https://www.kickstarter.com/activity")
-
-    self.vm.inputs.applicationDidFinishLaunching(application: .shared, launchOptions: [:])
-
-    self.goToActivity.assertValueCount(0)
-    XCTAssertFalse(self.vm.outputs.continueUserActivityReturnValue.value)
-
-    let result = self.vm.inputs.applicationContinueUserActivity(userActivity)
-    XCTAssertTrue(result)
-
-    self.goToActivity.assertValueCount(1)
-    XCTAssertTrue(self.vm.outputs.continueUserActivityReturnValue.value)
-  }
-
-  func testContinueUserActivity_InvalidActivity() {
-    let userActivity = NSUserActivity(activityType: "Other")
-
-    self.vm.inputs.applicationDidFinishLaunching(application: .shared, launchOptions: [:])
-    let result = self.vm.inputs.applicationContinueUserActivity(userActivity)
-    XCTAssertFalse(result)
-
-    XCTAssertFalse(self.vm.outputs.continueUserActivityReturnValue.value)
-  }
-
-  func testContinueUserActivity_Success() {
-    let userActivity = NSUserActivity(activityType: NSUserActivityTypeBrowsingWeb)
-    userActivity.webpageURL = URL(string: "https://www.kickstarter.com/activity")
-
-    withEnvironment(currentUser: nil) {
-      self.vm.inputs.applicationDidFinishLaunching(application: .shared, launchOptions: [:])
-
-      self.goToActivity.assertValueCount(0)
-      XCTAssertFalse(self.vm.outputs.continueUserActivityReturnValue.value)
-
-      let result = self.vm.inputs.applicationContinueUserActivity(userActivity)
-      XCTAssertTrue(result)
-
-      XCTAssertTrue(self.vm.outputs.continueUserActivityReturnValue.value)
-      self.goToActivity.assertValueCount(1)
-    }
-  }
-
   func testSetApplicationShortcutItems() {
     self.setApplicationShortcutItems.assertValues([])
 
@@ -1329,123 +664,6 @@ final class AppDelegateViewModelTests: TestCase {
       self.setApplicationShortcutItems.assertValues([
         [.recommendedForYou, .projectsWeLove, .search]
       ])
-    }
-  }
-
-  func testPerformShortcutItem_ProjectsWeLove() {
-    self.vm.inputs.applicationDidFinishLaunching(
-      application: UIApplication.shared,
-      launchOptions: [:]
-    )
-
-    self.goToDiscovery.assertValueCount(0)
-
-    self.vm.inputs.applicationPerformActionForShortcutItem(
-      ShortcutItem.projectsWeLove.applicationShortcutItem
-    )
-
-    let params = .defaults
-      |> DiscoveryParams.lens.staffPicks .~ true
-      |> DiscoveryParams.lens.sort .~ .magic
-    self.goToDiscovery.assertValues([params])
-  }
-
-  func testLaunchShortcutItem_ProjectsWeLove() {
-    self.vm.inputs.applicationDidFinishLaunching(
-      application: UIApplication.shared,
-      launchOptions: [
-        UIApplication.LaunchOptionsKey.shortcutItem: ShortcutItem.projectsWeLove.applicationShortcutItem
-      ]
-    )
-
-    let params = .defaults
-      |> DiscoveryParams.lens.staffPicks .~ true
-      |> DiscoveryParams.lens.sort .~ .magic
-    self.goToDiscovery.assertValues([params])
-    XCTAssertFalse(self.vm.outputs.applicationDidFinishLaunchingReturnValue)
-  }
-
-  func testPerformShortcutItem_RecommendedForYou() {
-    self.vm.inputs.applicationDidFinishLaunching(
-      application: UIApplication.shared,
-      launchOptions: [:]
-    )
-
-    self.goToDiscovery.assertValueCount(0)
-
-    self.vm.inputs.applicationPerformActionForShortcutItem(
-      ShortcutItem.recommendedForYou.applicationShortcutItem
-    )
-
-    let params = .defaults
-      |> DiscoveryParams.lens.recommended .~ true
-      |> DiscoveryParams.lens.sort .~ .magic
-    self.goToDiscovery.assertValues([params])
-  }
-
-  func testLaunchShortcutItem_RecommendedForYou() {
-    self.vm.inputs.applicationDidFinishLaunching(
-      application: UIApplication.shared,
-      launchOptions: [
-        UIApplication.LaunchOptionsKey.shortcutItem: ShortcutItem.recommendedForYou.applicationShortcutItem
-      ]
-    )
-
-    let params = .defaults
-      |> DiscoveryParams.lens.recommended .~ true
-      |> DiscoveryParams.lens.sort .~ .magic
-    self.goToDiscovery.assertValues([params])
-    XCTAssertFalse(self.vm.outputs.applicationDidFinishLaunchingReturnValue)
-  }
-
-  func testPerformShortcutItem_Search() {
-    self.vm.inputs.applicationDidFinishLaunching(
-      application: UIApplication.shared,
-      launchOptions: [:]
-    )
-
-    self.goToSearch.assertValueCount(0)
-
-    self.vm.inputs.applicationPerformActionForShortcutItem(ShortcutItem.search.applicationShortcutItem)
-
-    self.goToSearch.assertValueCount(1)
-  }
-
-  func testPerformShortcutItem_Success() {
-    withEnvironment(currentUser: nil) {
-      self.vm.inputs.applicationDidFinishLaunching(
-        application: UIApplication.shared,
-        launchOptions: [:]
-      )
-      self.vm.inputs.applicationPerformActionForShortcutItem(ShortcutItem.search.applicationShortcutItem)
-
-      self.goToSearch.assertValueCount(1)
-    }
-  }
-
-  func testLaunchShortcutItem_Search() {
-    self.vm.inputs.applicationDidFinishLaunching(
-      application: UIApplication.shared,
-      launchOptions: [
-        UIApplication.LaunchOptionsKey.shortcutItem: ShortcutItem.search.applicationShortcutItem
-      ]
-    )
-
-    self.goToSearch.assertValueCount(1)
-    XCTAssertFalse(self.vm.outputs.applicationDidFinishLaunchingReturnValue)
-  }
-
-  func testLaunchShortcutItem_Failure() {
-    withEnvironment(currentUser: nil) {
-      self.vm.inputs.applicationDidFinishLaunching(
-        application: UIApplication.shared,
-        launchOptions: [
-          UIApplication.LaunchOptionsKey.shortcutItem: ShortcutItem.search.applicationShortcutItem
-        ]
-      )
-
-      self.goToSearch.assertValueCount(1)
-      XCTAssertFalse(self.vm.outputs.applicationDidFinishLaunchingReturnValue)
     }
   }
 
@@ -1590,226 +808,6 @@ final class AppDelegateViewModelTests: TestCase {
     )
   }
 
-  func testEmailDeepLinking() {
-    withEnvironment(apiService: MockService(fetchProjectResult: .success(.template))) {
-      let emailUrl = URL(string: "https://clicks.kickstarter.com/?qs=deadbeef")!
-
-      // The application launches.
-      self.vm.inputs.applicationDidFinishLaunching(
-        application: UIApplication.shared,
-        launchOptions: [:]
-      )
-
-      self.findRedirectUrl.assertValues([])
-      self.presentViewController.assertValues([])
-      self.goToMobileSafari.assertValues([])
-
-      // We deep-link to an email url.
-      self.vm.inputs.applicationDidEnterBackground()
-      self.vm.inputs.applicationWillEnterForeground()
-      let result = self.vm.inputs.applicationOpenUrl(
-        application: UIApplication.shared,
-        url: emailUrl,
-        options: [:]
-      )
-      XCTAssertTrue(result)
-
-      self.findRedirectUrl.assertValues([emailUrl], "Ask to find the redirect after open the email url.")
-      self.presentViewController.assertValues([], "No view controller is presented yet.")
-      self.goToMobileSafari.assertValues([])
-
-      // We find the redirect to be a project url.
-      self.vm.inputs.foundRedirectUrl(URL(string: "https://www.kickstarter.com/projects/creator/project")!)
-
-      self.findRedirectUrl.assertValues([emailUrl], "Nothing new is emitted.")
-      self.presentViewController.assertValueCount(1, "Present the project view controller.")
-      self.goToMobileSafari.assertValues([])
-    }
-  }
-
-  func testEmailDeepLinking_ContinuedUserActivity() {
-    withEnvironment(apiService: MockService(fetchProjectResult: .success(.template))) {
-      let emailUrl = URL(string: "https://emails.kickstarter.com/?qs=deadbeef")!
-      let userActivity = NSUserActivity(activityType: NSUserActivityTypeBrowsingWeb)
-      userActivity.webpageURL = emailUrl
-
-      // The application launches.
-      self.vm.inputs.applicationDidFinishLaunching(
-        application: UIApplication.shared,
-        launchOptions: [:]
-      )
-
-      self.findRedirectUrl.assertValues([])
-      self.presentViewController.assertValues([])
-      self.goToMobileSafari.assertValues([])
-
-      // We deep-link to an email url.
-      self.vm.inputs.applicationDidEnterBackground()
-      self.vm.inputs.applicationWillEnterForeground()
-      let result = self.vm.inputs.applicationContinueUserActivity(userActivity)
-      XCTAssertTrue(result)
-
-      self.findRedirectUrl.assertValues([emailUrl], "Ask to find the redirect after open the email url.")
-      self.presentViewController.assertValues([], "No view controller is presented yet.")
-      self.goToMobileSafari.assertValues([])
-
-      // We find the redirect to be a project url.
-      self.vm.inputs.foundRedirectUrl(URL(string: "https://www.kickstarter.com/projects/creator/project")!)
-
-      self.findRedirectUrl.assertValues([emailUrl], "Nothing new is emitted.")
-      self.presentViewController.assertValueCount(1, "Present the project view controller.")
-      self.goToMobileSafari.assertValues([])
-    }
-  }
-
-  func testEmailDeepLinking_UnrecognizedUrl() {
-    let emailUrl = URL(string: "https://clicks.kickstarter.com/?qs=deadbeef")!
-
-    // The application launches.
-    self.vm.inputs.applicationDidFinishLaunching(
-      application: UIApplication.shared,
-      launchOptions: [:]
-    )
-
-    self.findRedirectUrl.assertValues([])
-    self.presentViewController.assertValueCount(0)
-    self.goToMobileSafari.assertValues([])
-
-    // We deep-link to an email url.
-    self.vm.inputs.applicationDidEnterBackground()
-    self.vm.inputs.applicationWillEnterForeground()
-    let result = self.vm.inputs.applicationOpenUrl(
-      application: UIApplication.shared,
-      url: emailUrl,
-      options: [:]
-    )
-    XCTAssertTrue(result)
-
-    self.findRedirectUrl.assertValues([emailUrl], "Ask to find the redirect after open the email url.")
-    self.presentViewController.assertValues([], "No view controller is presented.")
-    self.goToMobileSafari.assertValues([], "Do not go to mobile safari")
-
-    // We find the redirect to be an unrecognized url.
-    let unrecognizedUrl = URL(string: "https://www.kickstarter.com/unreconizable")!
-    self.vm.inputs.foundRedirectUrl(unrecognizedUrl)
-
-    self.findRedirectUrl.assertValues([emailUrl], "Nothing new is emitted.")
-    self.presentViewController.assertValues([], "Do not present controller since the url was unrecognizable.")
-    self.goToMobileSafari.assertValues([unrecognizedUrl], "Go to mobile safari for the unrecognized url.")
-  }
-
-  func testEmailDeepLinking_UnrecognizedUrl_ProjectPreview() {
-    let emailUrl = URL(string: "https://emails.kickstarter.com/?qs=deadbeef")!
-
-    // The application launches.
-    self.vm.inputs.applicationDidFinishLaunching(
-      application: UIApplication.shared,
-      launchOptions: [:]
-    )
-
-    self.findRedirectUrl.assertValues([])
-    self.presentViewController.assertValueCount(0)
-    self.goToMobileSafari.assertValues([])
-
-    // We deep-link to an email url.
-    self.vm.inputs.applicationDidEnterBackground()
-    self.vm.inputs.applicationWillEnterForeground()
-    let result = self.vm.inputs.applicationOpenUrl(
-      application: UIApplication.shared,
-      url: emailUrl,
-      options: [:]
-    )
-    XCTAssertTrue(result)
-
-    self.findRedirectUrl.assertValues([emailUrl], "Ask to find the redirect after open the email url.")
-    self.presentViewController.assertValues([], "No view controller is presented.")
-    self.goToMobileSafari.assertValues([], "Do not go to mobile safari")
-
-    // We find the redirect to be an unrecognized url (project preview).
-    let unrecognizedUrl = URL(string: "https://www.kickstarter.com/projects/creator/project?token=4")!
-    self.vm.inputs.foundRedirectUrl(unrecognizedUrl)
-
-    self.findRedirectUrl.assertValues([emailUrl], "Nothing new is emitted.")
-    self.presentViewController.assertValues([], "Do not present controller since the url was unrecognizable.")
-    self.goToMobileSafari.assertValues([unrecognizedUrl], "Go to mobile safari for the unrecognized url.")
-  }
-
-  func testProjectSurveyDeepLink() {
-    self.vm.inputs.applicationDidFinishLaunching(
-      application: UIApplication.shared,
-      launchOptions: [:]
-    )
-
-    self.presentViewController.assertValues([])
-
-    let projectUrl = "https://www.kickstarter.com"
-      + "/projects/tequila/help-me-transform-this-pile-of-wood/surveys/123"
-    let result = self.vm.inputs.applicationOpenUrl(
-      application: UIApplication.shared,
-      url: URL(string: projectUrl)!,
-      options: [:]
-    )
-    XCTAssertTrue(result)
-
-    self.presentViewController.assertValues([1])
-  }
-
-  func testErroredPledgeDeepLink_LoggedIn() {
-    let project = Project.template
-      |> \.personalization.backing .~ .template
-    let service = MockService(fetchProjectResult: .success(project))
-
-    withEnvironment(apiService: service, currentUser: .template) {
-      self.vm.inputs.applicationDidFinishLaunching(
-        application: UIApplication.shared,
-        launchOptions: [:]
-      )
-
-      self.goToLoginWithIntent.assertDidNotEmitValue()
-      self.presentViewController.assertValues([])
-
-      let projectUrl = "https://www.kickstarter.com"
-        + "/projects/sshults/greensens-the-easy-way-to-take-care-of-your-houseplants-0"
-        + "/pledge?at=4f7d35e7c9d2bb57&ref=ksr_email_backer_failed_transaction"
-
-      let result = self.vm.inputs.applicationOpenUrl(
-        application: UIApplication.shared,
-        url: URL(string: projectUrl)!,
-        options: [:]
-      )
-      XCTAssertTrue(result)
-
-      self.goToLoginWithIntent.assertDidNotEmitValue()
-      self.presentViewController.assertValues([2])
-    }
-  }
-
-  func testErroredPledgeDeepLink_LoggedOut() {
-    withEnvironment(apiService: MockService(fetchProjectResult: .success(.template)), currentUser: nil) {
-      self.vm.inputs.applicationDidFinishLaunching(
-        application: UIApplication.shared,
-        launchOptions: [:]
-      )
-
-      self.goToLoginWithIntent.assertDidNotEmitValue()
-      self.presentViewController.assertValues([])
-
-      let projectUrl = "https://www.kickstarter.com"
-        + "/projects/sshults/greensens-the-easy-way-to-take-care-of-your-houseplants-0"
-        + "/pledge?at=4f7d35e7c9d2bb57&ref=ksr_email_backer_failed_transaction"
-
-      let result = self.vm.inputs.applicationOpenUrl(
-        application: UIApplication.shared,
-        url: URL(string: projectUrl)!,
-        options: [:]
-      )
-      XCTAssertTrue(result)
-
-      self.goToLoginWithIntent.assertValues([.erroredPledge])
-      self.presentViewController.assertDidNotEmitValue()
-    }
-  }
-
   func testErroredPledgePushDeepLink_LoggedIn() {
     let project = Project.template
       |> \.personalization.backing .~ .template
@@ -1865,25 +863,6 @@ final class AppDelegateViewModelTests: TestCase {
     XCTAssertEqual(self.goToLoginWithIntent.values, [.generic, .onboarding])
   }
 
-  func testUserSurveyDeepLink() {
-    self.vm.inputs.applicationDidFinishLaunching(
-      application: UIApplication.shared,
-      launchOptions: [:]
-    )
-
-    self.presentViewController.assertValues([])
-
-    let projectUrl = "https://www.kickstarter.com/users/tequila/surveys/123"
-    let result = self.vm.inputs.applicationOpenUrl(
-      application: UIApplication.shared,
-      url: URL(string: projectUrl)!,
-      options: [:]
-    )
-    XCTAssertTrue(result)
-
-    self.presentViewController.assertValues([1])
-  }
-
   func testShowAlertEmitsIf_CanShowDialog() {
     let notification = Notification(
       name: Notification.Name(rawValue: "deadbeef"),
@@ -1920,105 +899,6 @@ final class AppDelegateViewModelTests: TestCase {
       self.vm.inputs.showNotificationDialog(notification: notification)
 
       self.showAlert.assertDidNotEmitValue()
-    }
-  }
-
-  func testVerifyEmail_Success() {
-    self.emailVerificationCompletedMessage.assertDidNotEmitValue()
-    self.emailVerificationCompletedSuccess.assertDidNotEmitValue()
-
-    guard let url = URL(string: "https://www.kickstarter.com/profile/verify_email?at=12345") else {
-      XCTFail("Should have a url")
-      return
-    }
-
-    let env = EmailVerificationResponseEnvelope(
-      message: "Thanks—you’ve successfully verified your email address."
-    )
-
-    let mockService = MockService(verifyEmailResult: .success(env))
-
-    withEnvironment(apiService: mockService) {
-      _ = self.vm.inputs.applicationOpenUrl(
-        application: UIApplication.shared,
-        url: url,
-        options: [:]
-      )
-
-      self.scheduler.advance()
-
-      self.emailVerificationCompletedSuccess.assertValues([true])
-      self.emailVerificationCompletedMessage.assertValues(
-        ["Thanks—you’ve successfully verified your email address."]
-      )
-    }
-  }
-
-  func testVerifyEmail_Failure() {
-    self.emailVerificationCompletedMessage.assertDidNotEmitValue()
-    self.emailVerificationCompletedSuccess.assertDidNotEmitValue()
-
-    guard let url = URL(string: "https://www.kickstarter.com/profile/verify_email?at=12345") else {
-      XCTFail("Should have a url")
-      return
-    }
-
-    let errorEnvelope = ErrorEnvelope(
-      errorMessages: ["Error Message"],
-      ksrCode: .UnknownCode,
-      httpCode: 403,
-      exception: nil
-    )
-
-    let mockService = MockService(verifyEmailResult: .failure(errorEnvelope))
-
-    withEnvironment(apiService: mockService) {
-      _ = self.vm.inputs.applicationOpenUrl(
-        application: UIApplication.shared,
-        url: url,
-        options: [:]
-      )
-
-      self.scheduler.advance()
-
-      self.emailVerificationCompletedSuccess.assertValues([false])
-      self.emailVerificationCompletedMessage.assertValues(
-        ["Error Message"]
-      )
-    }
-  }
-
-  func testVerifyEmail_Failure_UnknownError() {
-    self.emailVerificationCompletedMessage.assertDidNotEmitValue()
-    self.emailVerificationCompletedSuccess.assertDidNotEmitValue()
-
-    guard let url = URL(string: "https://www.kickstarter.com/profile/verify_email?at=12345") else {
-      XCTFail("Should have a url")
-      return
-    }
-
-    let errorEnvelope = ErrorEnvelope(
-      errorMessages: [],
-      ksrCode: .UnknownCode,
-      httpCode: 500,
-      exception: nil
-    )
-
-    let mockService = MockService(verifyEmailResult: .failure(errorEnvelope))
-
-    withEnvironment(apiService: mockService) {
-      _ = self.vm.inputs.applicationOpenUrl(
-        application: UIApplication.shared,
-        url: url,
-        options: [:]
-      )
-
-      self.scheduler.advance()
-
-      self.emailVerificationCompletedSuccess.assertValues([false])
-      self.emailVerificationCompletedMessage.assertValues(
-        ["Something went wrong, please try again."]
-      )
     }
   }
 
@@ -2155,46 +1035,6 @@ final class AppDelegateViewModelTests: TestCase {
       self.vm.inputs.configUpdatedNotificationObserved()
 
       self.segmentIsEnabled.assertValues([false])
-    }
-  }
-
-  func testDeepLink_UserDidUpdateNotificationSettings() {
-    self.updateCurrentUserInEnvironment.assertDidNotEmitValue()
-
-    withEnvironment(apiService: MockService()) {
-      let user = User.template
-        |> User.lens.notifications.mobileMessages .~ false
-
-      let env = AccessTokenEnvelope(accessToken: "deadbeef", user: user)
-      AppEnvironment.login(env)
-
-      self.vm.inputs.applicationDidFinishLaunching(
-        application: UIApplication.shared,
-        launchOptions: [:]
-      )
-
-      self.scheduler.advance(by: .seconds(5))
-
-      self.updateCurrentUserInEnvironment.assertValues([user])
-
-      let updatedUser = user
-        |> User.lens.notifications.mobileMessages .~ true
-
-      let url =
-        "https://\(AppEnvironment.current.apiService.serverConfig.webBaseUrl.host ?? "")/settings/notify_mobile_of_messages/true"
-
-      let result = self.vm.inputs.applicationOpenUrl(
-        application: UIApplication.shared,
-        url: URL(string: url)!,
-        options: [:]
-      )
-      XCTAssertTrue(result)
-
-      self.updateCurrentUserInEnvironment.assertValues([user])
-
-      self.scheduler.advance()
-
-      self.updateCurrentUserInEnvironment.assertValues([user, updatedUser])
     }
   }
 
@@ -2571,54 +1411,6 @@ final class AppDelegateViewModelTests: TestCase {
 
     self.goToMobileSafari.assertValues([url])
     self.presentViewController.assertValues([])
-  }
-
-  func testGoToMobileSafari_unrecognizedDeeplink() {
-    self.vm.inputs.applicationDidFinishLaunching(
-      application: UIApplication.shared,
-      launchOptions: [:]
-    )
-
-    let url = URL(string: "https://fake-url.com")!
-    _ = self.vm.inputs.applicationOpenUrl(application: nil, url: url, options: [:])
-
-    self.goToMobileSafari.assertLastValue(url)
-  }
-
-  func testGoToMobileSafari_deeplinkFound() {
-    self.vm.inputs.applicationDidFinishLaunching(
-      application: UIApplication.shared,
-      launchOptions: [:]
-    )
-
-    let url = URL(string: "https://kickstarter.com/activity")!
-    _ = self.vm.inputs.applicationOpenUrl(application: nil, url: url, options: [:])
-
-    self.goToMobileSafari.assertDidNotEmitValue()
-  }
-
-  func testGoToMobileSafari_ksrUnrecognizedDeeplink() throws {
-    self.vm.inputs.applicationDidFinishLaunching(
-      application: UIApplication.shared,
-      launchOptions: [:]
-    )
-
-    let url = URL(string: "ksr://fake-url.com")!
-    _ = self.vm.inputs.applicationOpenUrl(application: nil, url: url, options: [:])
-
-    self.goToMobileSafari.assertDidNotEmitValue()
-  }
-
-  func testGoToMobileSafari_ksrDeeplinkFound() {
-    self.vm.inputs.applicationDidFinishLaunching(
-      application: UIApplication.shared,
-      launchOptions: [:]
-    )
-
-    let url = URL(string: "https://kickstarter.com/activity")!
-    _ = self.vm.inputs.applicationOpenUrl(application: nil, url: url, options: [:])
-
-    self.goToMobileSafari.assertDidNotEmitValue()
   }
 
   func testRemoteConfigClientConfiguredNotification_Success() {
